Обновления для версии 1.0.0
