+++ index.html
@@ -13,6 +13,6 @@
       <!-- <p id="author">(author)</p> -->
       <button id="generate-btn">Generate Quote</button>
     </div>
-    <script src="script.js"></script>
+    <script src="script.js" type="module"></script>
   </body>
 </html>
