+++ script.js
@@ -1,45 +1,5 @@
-const quotes = [
-    {
-      text: 'The only way to do great work is to love what you do.',
-      author: 'Steve Jobs',
-    },
-    {
-      text: 'Innovation distinguishes between a leader and a follower.',
-      author: 'Steve Jobs',
-    },
-    {
-      text: "Your time is limited, so don't waste it living someone else's life.",
-      author: 'Steve Jobs',
-    },
-    {
-      text: 'Stay hungry, stay foolish.',
-      author: 'Steve Jobs',
-    },
-    {
-      text: 'The people who are crazy enough to think they can change the world are the ones who do.',
-      author: 'Steve Jobs',
-    },
-    {
-      text: 'Sometimes life hits you in the head with a brick. Don’t lose faith.',
-      author: 'Steve Jobs',
-    },
-    {
-      text: 'I’m convinced that about half of what separates successful entrepreneurs from the non-successful ones is pure perseverance.',
-      author: 'Steve Jobs',
-    },
-    {
-      text: 'Great things in business are never done by one person. They’re done by a team of people.',
-      author: 'Steve Jobs',
-    },
-    {
-      text: 'Be a yardstick of quality. Some people aren’t used to an environment where excellence is expected.',
-      author: 'Steve Jobs',
-    },
-    {
-      text: 'My favorite things in life don’t cost any money. It’s really clear that the most precious resource we all have is time.',
-      author: 'Steve Jobs',
-    }
-  ];
+import quotes from "./modules/quotes.js";
+
   
   const quoteElement = document.getElementById('quote');
   const generateBtn = document.getElementById('generate-btn');
